requirements.txt
alembic==1.14.0
anyio==4.6.2.post1
blinker==1.9.0
cerbos==0.11.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
mypy-extensions==1.0.0
packaging==24.2
protobuf==5.28.3
protoc-gen-openapiv2==0.0.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
types-protobuf==5.28.3.20241030
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
